requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
openai==1.51.0
google-generativeai==0.8.3
qdrant-client==1.15.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
numpy==1.25.2
pandas==2.1.3
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
geopy==2.4.1
folium==0.15.0
bcrypt==4.0.1
pyjwt==2.8.0